:sparkles: Share values between components (#545)
